requirements.txt
firebase-admin
opencv-python
numpy
torch
runpod
pillow
supervision